functions/requirements.txt
firebase_functions==0.1.0
firebase-admin==6.5.0
pydantic==2.9.2
openai==1.39.0
requests==2.32.3
aiohttp==3.9.5
google-cloud-storage==2.18.2
google-cloud-firestore==2.16.0
google-cloud-logging==3.11.3
python-dotenv==1.0.1
weaviate-client==4.18.0
